Open sidebar when highlighting while logged-out
When the user creates a highlight while logged out, they appear to
work however, they silently do not get saved. An error message
does pop up in the sidebar informing them they cannot create annotations
when not logged in but they don't see it because the sidebar does not
open when creating a highlight. This makes the error message more
apparent to the user by popping open the sidebar.